Update GawaGo dashboard and worker UI

+++ backend/apps/reviews/models.py
@@ -15,7 +15,7 @@ class Review(models.Model):
     author_role = models.CharField(max_length=20, choices=AUTHOR_ROLE_CHOICES)
     target_role = models.CharField(max_length=20, choices=AUTHOR_ROLE_CHOICES)
     job_title = models.CharField(max_length=255, blank=True, default="")
-    rating = models.PositiveSmallIntegerField(null=True, blank=True)
+    rating = models.DecimalField(max_digits=2, decimal_places=1, null=True, blank=True)
     feedback = models.TextField(blank=True, default="")
     created_at = models.DateTimeField(auto_now_add=True)
